requirements.txt

dash==2.14.2
dash-bootstrap-components==1.5.0
plotly==5.18.0
pandas==2.1.4
mysql-connector-python==8.2.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
gunicorn==20.1.0
